+++ test/Spec.hs
@@ -2,6 +2,7 @@ import           Test.Hspec
 
 import           Test.GetAllMetrics
 import           Test.GetMetrics
+import           Test.GetUpdatedMetrics
 
 main :: IO ()
 main = hspec $ do
@@ -10,8 +11,12 @@ main = hspec $ do
       getAllMetricsViaPipe
     it "request of some metrics"
       getMetricsViaPipe
+    it "request of updated metrics"
+      getUpdatedMetricsViaPipe
   describe "EKG metrics forwarding, via remote socket" $ do
     it "request of all metrics"
       getAllMetricsViaSocket
     it "request of some metrics"
       getMetricsViaSocket
+    it "request of updated metrics"
+      getUpdatedMetricsViaSocket